# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for HealthPoints clinic management platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Clinic {
  id: number;
  name: string;
  nameAr: string;
  specialty: string;
  specialtyAr: string;
  phone: string;
  email: string;
  address: string;
  addressAr: string;
  city: string;
  openHours: string;
  rating: number;
  isActive: boolean;
  totalPatients: number;
  pointsPerVisit: number;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  descriptionAr?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
}

export interface ClinicInput {
  name: string;
  nameAr: string;
  specialty: string;
  specialtyAr: string;
  phone: string;
  email: string;
  address: string;
  addressAr: string;
  city: string;
  openHours: string;
  pointsPerVisit: number;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  descriptionAr?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface Offer {
  id: number;
  clinicId: number;
  clinicName: string;
  title: string;
  titleAr: string;
  description: string;
  descriptionAr: string;
  originalPrice: number;
  discountedPrice: number;
  pointsRequired: number;
  category: string;
  isFeatured: boolean;
  isActive: boolean;
  expiryDate: string;
  usageCount: number;
  createdAt: string;
}

export interface OfferInput {
  clinicId: number;
  title: string;
  titleAr: string;
  description: string;
  descriptionAr: string;
  originalPrice: number;
  discountedPrice: number;
  pointsRequired: number;
  category: string;
  isFeatured: boolean;
  isActive: boolean;
  expiryDate: string;
}

export type AppointmentStatus =
  (typeof AppointmentStatus)[keyof typeof AppointmentStatus];

export const AppointmentStatus = {
  pending: "pending",
  confirmed: "confirmed",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Appointment {
  id: number;
  patientName: string;
  patientPhone: string;
  clinicId: number;
  clinicName: string;
  service: string;
  serviceAr: string;
  date: string;
  time: string;
  status: AppointmentStatus;
  pointsEarned: number;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export type DiscountRequiredLevel =
  (typeof DiscountRequiredLevel)[keyof typeof DiscountRequiredLevel];

export const DiscountRequiredLevel = {
  bronze: "bronze",
  silver: "silver",
  gold: "gold",
  platinum: "platinum",
} as const;

export interface Discount {
  id: number;
  code: string;
  title: string;
  titleAr: string;
  description: string;
  discountPercent: number;
  requiredLevel: DiscountRequiredLevel;
  /** @nullable */
  clinicId?: number | null;
  /** @nullable */
  clinicName?: string | null;
  service: string;
  expiryDate: string;
  isActive: boolean;
  usageCount: number;
  maxUsage: number;
  createdAt: string;
}

export type DiscountInputRequiredLevel =
  (typeof DiscountInputRequiredLevel)[keyof typeof DiscountInputRequiredLevel];

export const DiscountInputRequiredLevel = {
  bronze: "bronze",
  silver: "silver",
  gold: "gold",
  platinum: "platinum",
} as const;

export interface DiscountInput {
  code: string;
  title: string;
  titleAr: string;
  description: string;
  discountPercent: number;
  requiredLevel: DiscountInputRequiredLevel;
  /** @nullable */
  clinicId?: number | null;
  service: string;
  expiryDate: string;
  isActive: boolean;
  maxUsage: number;
}

export type DashboardStatsLevelDistribution = {
  bronze: number;
  silver: number;
  gold: number;
  platinum: number;
};

export type DashboardStatsMonthlyAppointmentsItem = {
  month: string;
  count: number;
};

export interface DashboardStats {
  totalPatients: number;
  totalClinics: number;
  totalAppointments: number;
  totalPointsIssued: number;
  activeOffers: number;
  revenueThisMonth: number;
  appointmentsThisMonth: number;
  newPatientsThisMonth: number;
  levelDistribution: DashboardStatsLevelDistribution;
  monthlyAppointments: DashboardStatsMonthlyAppointmentsItem[];
}

export type PatientLevel = (typeof PatientLevel)[keyof typeof PatientLevel];

export const PatientLevel = {
  bronze: "bronze",
  silver: "silver",
  gold: "gold",
  platinum: "platinum",
} as const;

export interface Patient {
  id: number;
  name: string;
  phone: string;
  email: string;
  level: PatientLevel;
  points: number;
  totalVisits: number;
  joinedAt: string;
  /** @nullable */
  lastVisit?: string | null;
}

export type UpdateAppointmentStatusBodyStatus =
  (typeof UpdateAppointmentStatusBodyStatus)[keyof typeof UpdateAppointmentStatusBodyStatus];

export const UpdateAppointmentStatusBodyStatus = {
  pending: "pending",
  confirmed: "confirmed",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type UpdateAppointmentStatusBody = {
  status: UpdateAppointmentStatusBodyStatus;
};