# lib/api-zod/src/generated/types/clinicInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for HealthPoints clinic management platform
 * OpenAPI spec version: 0.1.0
 */

export interface ClinicInput {
  name: string;
  nameAr: string;
  specialty: string;
  specialtyAr: string;
  phone: string;
  email: string;
  address: string;
  addressAr: string;
  city: string;
  openHours: string;
  pointsPerVisit: number;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  descriptionAr?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}