# lib/api-zod/src/generated/types/discountInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for HealthPoints clinic management platform
 * OpenAPI spec version: 0.1.0
 */
import type { DiscountInputRequiredLevel } from "./discountInputRequiredLevel";

export interface DiscountInput {
  code: string;
  title: string;
  titleAr: string;
  description: string;
  discountPercent: number;
  requiredLevel: DiscountInputRequiredLevel;
  /** @nullable */
  clinicId?: number | null;
  service: string;
  expiryDate: string;
  isActive: boolean;
  maxUsage: number;
}