# lib/api-zod/src/generated/types/offer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for HealthPoints clinic management platform
 * OpenAPI spec version: 0.1.0
 */

export interface Offer {
  id: number;
  clinicId: number;
  clinicName: string;
  title: string;
  titleAr: string;
  description: string;
  descriptionAr: string;
  originalPrice: number;
  discountedPrice: number;
  pointsRequired: number;
  category: string;
  isFeatured: boolean;
  isActive: boolean;
  expiryDate: string;
  usageCount: number;
  createdAt: string;
}