# lib/api-zod/src/generated/types/patient.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for HealthPoints clinic management platform
 * OpenAPI spec version: 0.1.0
 */
import type { PatientLevel } from "./patientLevel";

export interface Patient {
  id: number;
  name: string;
  phone: string;
  email: string;
  level: PatientLevel;
  points: number;
  totalVisits: number;
  joinedAt: string;
  /** @nullable */
  lastVisit?: string | null;
}